// frontend/src/api/models/Canvas.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type Canvas = {
    id: number;
    user_id: number;
    name?: string | null;
    memo?: string | null;
    width?: number | null;
    height?: number | null;
    created_at: string;
    updated_at: string;
};