// frontend/src/api/models/CanvasImage.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CanvasImage = {
    id: number;
    canvas_id: number;
    uri: string;
    'x'?: number | null;
    'y'?: number | null;
    width?: number | null;
    height?: number | null;
    left?: number | null;
    right?: number | null;
    top?: number | null;
    bottom?: number | null;
    size?: number | null;
    created_at: string;
    updated_at: string;
};